kernel-5.15: fix spec file
In 5b853d34 the 5.15 kernel spec file was not updated to extract the
kernel.spec file from the SRPM, preventing the patches from being
applied. This update extracts the kernel.spec file so the patch command
works as expected.

Additionally, if the number of patch files in the SRPM doesn't match the
number of patches found, fail the build.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>